move exhaustive search to qt
